refactored UserMovieLists to return an error message if the list is empty, fixed the trailer undefined error
